Simplify Process methods
- The first frame is already passed to the writers, allowing us to wait until the frame is passed instead of requiring a Replay/Select/SelectMany pattern to build the expression trees
