+++ OpenEphys.Onix1/DataFrameWriter/BufferedDataFrameArrowFileSink.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using Apache.Arrow;
 using Bonsai.IO;
@@ -7,25 +7,21 @@ namespace OpenEphys.Onix1.DataFrameWriter
 {
     class BufferedDataFrameArrowFileSink : FileSink<BufferedDataFrame, ArrowBatchWriter<BufferedDataFrame>>
     {
-        readonly Schema schema;
-        readonly Func<IList<BufferedDataFrame>, Schema, RecordBatch> createRecordBatch;
-        readonly int bufferSize;
         readonly TimeSpan timeout;
 
-        public BufferedDataFrameArrowFileSink(
-            Schema schema,
-            Func<IList<BufferedDataFrame>, Schema, RecordBatch> createRecordBatch,
-            int bufferSize,
-            TimeSpan timeout)
+        public BufferedDataFrameArrowFileSink(TimeSpan timeout)
         {
-            this.schema = schema;
-            this.createRecordBatch = createRecordBatch;
-            this.bufferSize = bufferSize;
             this.timeout = timeout;
         }
 
         protected override ArrowBatchWriter<BufferedDataFrame> CreateWriter(string filename, BufferedDataFrame dataFrame)
         {
+            var frameType = dataFrame.GetType();
+            var members = DataFrameWriterHelper.GetDataMembers(frameType);
+            var schema = DataFrameWriterHelper.GenerateSchema(members, dataFrame);
+            var createRecordBatch = RecordBatchExpressionFactory.CreateBuilder<Func<IList<BufferedDataFrame>, Schema, RecordBatch>>(
+                new BufferedDataFrameExpressionProvider(), frameType, members).Compile();
+            var bufferSize = (int)Math.Ceiling((double)DataFrameWriterHelper.GetBufferSize(frameType) / dataFrame.Clock.Length);
             return new ArrowBatchWriter<BufferedDataFrame>(filename, schema, bufferSize, timeout, createRecordBatch);
         }
