+++ OpenEphys.Onix1/DataFrameWriter/DataFrameArrowFileSink.cs
@@ -7,25 +7,20 @@ namespace OpenEphys.Onix1.DataFrameWriter
 {
     class DataFrameArrowFileSink : FileSink<DataFrame, ArrowBatchWriter<DataFrame>>
     {
-        readonly Schema schema;
-        readonly Func<IList<DataFrame>, Schema, RecordBatch> createRecordBatch;
-        readonly int bufferSize;
         readonly TimeSpan timeout;
 
-        public DataFrameArrowFileSink(
-            Schema schema,
-            Func<IList<DataFrame>, Schema, RecordBatch> createRecordBatch,
-            int bufferSize,
-            TimeSpan timeout)
+        public DataFrameArrowFileSink(TimeSpan timeout)
         {
-            this.schema = schema;
-            this.createRecordBatch = createRecordBatch;
-            this.bufferSize = bufferSize;
             this.timeout = timeout;
         }
 
         protected override ArrowBatchWriter<DataFrame> CreateWriter(string filename, DataFrame frame)
         {
+            var frameType = frame.GetType();
+            var members = DataFrameWriterHelper.GetDataMembers(frameType);
+            var schema = DataFrameWriterHelper.GenerateSchema(members, frame);
+            var createRecordBatch = RecordBatchExpressionFactory.CreateBuilder<Func<IList<DataFrame>, Schema, RecordBatch>>( new DataFrameExpressionProvider(), frameType, members).Compile();
+            var bufferSize = DataFrameWriterHelper.GetBufferSize(frameType);
             return new ArrowBatchWriter<DataFrame>(filename, schema, bufferSize, timeout, createRecordBatch);
         }
