+++ OpenEphys.Onix1/DataFrameWriter/DataFrameWriter.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Reactive.Linq;
-using System.Reflection;
-using Apache.Arrow;
+using System;
 using Bonsai;
 using Bonsai.IO;
 
@@ -17,86 +11,7 @@ namespace OpenEphys.Onix1.DataFrameWriter
     [WorkflowElementCategory(ElementCategory.Sink)]
     public class DataFrameWriter : FileSink
     {
-        /// <summary>
-        /// Represents the number of seconds to wait before data is flushed, regardless of if <see
-        /// cref="MinimumBufferSize"/> samples have been collected.
-        /// </summary>
         const int SecondsBeforeFlush = 5;
-
-        /// <summary>
-        /// Represents the default number of <see cref="DataFrame">DataFrames</see> in a <see
-        /// cref="RecordBatch"/>.
-        /// </summary>
-        const int DefaultBufferSize = 1000;
-
-        /// <summary>
-        /// Represents the mimumum number of <see cref="DataFrame">DataFrames</see> in a <see
-        /// cref="RecordBatch"/>.
-        /// </summary>
-        const int MinimumBufferSize = 100;
-
-        BufferedDataFrameArrowFileSink CreateBufferedDataFrameArrowFileSink(
-            Schema schema,
-            Func<IList<BufferedDataFrame>, Schema, RecordBatch> createRecordBatch,
-            int bufferSize,
-            TimeSpan timeout)
-        {
-            return new BufferedDataFrameArrowFileSink(schema, createRecordBatch, bufferSize, timeout)
-            {
-                FileName = this.FileName,
-                Suffix = this.Suffix,
-                Buffered = this.Buffered,
-                Overwrite = this.Overwrite
-            };
-        }
-
-        DataFrameArrowFileSink CreateDataFrameArrowFileSink(
-            Schema schema,
-            Func<IList<DataFrame>, Schema, RecordBatch> createRecordBatch,
-            int bufferSize,
-            TimeSpan timeout)
-        {
-            return new DataFrameArrowFileSink(schema, createRecordBatch, bufferSize, timeout)
-            {
-                FileName = this.FileName,
-                Suffix = this.Suffix,
-                Buffered = this.Buffered,
-                Overwrite = this.Overwrite
-            };
-        }
-
-        static Expression<Func<IList<DataFrame>, Schema, RecordBatch>> CreateDataFrameRecordBatchBuilder(
-            Type frameType,
-            IEnumerable<MemberInfo> members)
-        {
-            return RecordBatchExpressionFactory.CreateBuilder<Func<IList<DataFrame>, Schema, RecordBatch>>(
-                new DataFrameExpressionProvider(),
-                frameType,
-                members);
-        }
-
-        static Expression<Func<IList<BufferedDataFrame>, Schema, RecordBatch>> CreateBufferedFrameRecordBatchBuilder(
-            Type frameType,
-            IEnumerable<MemberInfo> members)
-        {
-            return RecordBatchExpressionFactory.CreateBuilder<Func<IList<BufferedDataFrame>, Schema, RecordBatch>>(
-                new BufferedDataFrameExpressionProvider(),
-                frameType,
-                members);
-        }
-
-        static int GetBufferSize(Type frameType)
-        {
-            var sampleRateAttribute = frameType.GetCustomAttribute<ExpectedSampleRateAttribute>();
-            if (sampleRateAttribute != null)
-            {
-                const double BufferDurationSeconds = 1.0;
-                var bufferSize = (int)(sampleRateAttribute.SampleRateHz * BufferDurationSeconds);
-                return bufferSize >= MinimumBufferSize ? bufferSize : MinimumBufferSize;
-            }
-
-            return DefaultBufferSize;
-        }
 
         /// <summary>
         /// Writes all of the data frames in the sequence to an Apache Arrow file.
@@ -108,21 +23,13 @@ namespace OpenEphys.Onix1.DataFrameWriter
         /// </returns>
         public IObservable<BufferedDataFrame> Process(IObservable<BufferedDataFrame> source)
         {
-            return source.Replay(frames =>
+            return new BufferedDataFrameArrowFileSink(TimeSpan.FromSeconds(SecondsBeforeFlush))
             {
-                var config = frames.Take(1).Select(frame =>
-                {
-                    var frameType = frame.GetType();
-                    var members = DataFrameWriterHelper.GetDataMembers(frameType);
-                    var bufferSize = (int)Math.Ceiling((double)GetBufferSize(frameType) / frame.Clock.Length);
-                    var schema = DataFrameWriterHelper.GenerateSchema(members, frame);
-                    var createRecordBatch = CreateBufferedFrameRecordBatchBuilder(frameType, members).Compile();
-                    return (schema, createRecordBatch, bufferSize);
-                });
-
-                return config.SelectMany(c => CreateBufferedDataFrameArrowFileSink(c.schema, c.createRecordBatch, c.bufferSize, 
-                    TimeSpan.FromSeconds(SecondsBeforeFlush)).Process(frames));
-            }, 1);
+                FileName = this.FileName,
+                Suffix = this.Suffix,
+                Buffered = this.Buffered,
+                Overwrite = this.Overwrite
+            }.Process(source);
         }
 
         /// <summary>
@@ -135,21 +42,13 @@ namespace OpenEphys.Onix1.DataFrameWriter
         /// </returns>
         public IObservable<DataFrame> Process(IObservable<DataFrame> source)
         {
-            return source.Replay(frames =>
+            return new DataFrameArrowFileSink(TimeSpan.FromSeconds(SecondsBeforeFlush))
             {
-                var config = frames.Take(1).Select(frame =>
-                {
-                    var frameType = frame.GetType();
-                    var members = DataFrameWriterHelper.GetDataMembers(frameType);
-                    var bufferSize = GetBufferSize(frameType);
-                    var schema = DataFrameWriterHelper.GenerateSchema(members, frame);
-                    var createRecordBatch = CreateDataFrameRecordBatchBuilder(frameType, members).Compile();
-                    return (schema, createRecordBatch, bufferSize);
-                });
-
-                return config.SelectMany(c => CreateDataFrameArrowFileSink(c.schema, c.createRecordBatch, c.bufferSize, 
-                    TimeSpan.FromSeconds(SecondsBeforeFlush)).Process(frames));
-            }, 1);
+                FileName = this.FileName,
+                Suffix = this.Suffix,
+                Buffered = this.Buffered,
+                Overwrite = this.Overwrite
+            }.Process(source);
         }
     }
 }
